refactored courses
