+++ server/__init__.py
@@ -34,7 +34,8 @@ CORS(app)
 db = redis.StrictRedis(host='localhost', port=6379, db=0)
 app.secret_key = os.urandom(24)
 
-import server.account  # noqa
+import server.account.account  # noqa
+import server.account.settings  # noqa
 import server.analytics  # noqa
 import server.auth  # noqa
 import server.buildings  # noqa
