+++ server/account/settings.py
@@ -1,0 +1,17 @@
+from flask import g, jsonify, request
+
+from server import app
+from server.auth import auth
+from server.notifications import get_notification_settings
+from server.privacy import get_privacy_settings
+
+
+@app.route('/account/settings', methods=['GET'])
+@auth()
+def get_account_settings():
+    notifSettings = get_notification_settings(g.account)
+    privacySettings = get_privacy_settings(g.account)
+    return jsonify({
+    	'notification_settings': notifSettings,
+    	'privacy_settings': privacySettings,
+    })
